pyproject.toml
[project]
name = "simhops"
version = "0.1.0"
description = "3D quadcopter simulator with Gymnasium RL environment"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "gymnasium>=0.29.0",
    "mujoco>=3.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "rerun-sdk>=0.18.0",
    "rich>=13.0.0",
    "stable-baselines3>=2.2.0",
    "tensorboard>=2.14.0",
    "torch>=2.0.0",
    "tqdm>=4.65.0",
]

[project.scripts]
simhops-train = "simhops.train:main"
simhops-eval = "simhops.evaluate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/simhops"]
